fix: resolve staged user_custom in Setup for remote subprocesses
